About code
This code finds the value of sine by using the expression derived by the Taylor series expansion method. It is recommended to take as large value of k as possible. It is so as the addition in the method for sine determination can go up to infinity. So, more the terms added, more is the precision. Therefore, larger the value of k, better the result.